% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aqtrends_helpers.R
\name{average_data}
\alias{average_data}
\title{Average raw hourly data by year or month-year}
\usage{
average_data(df, avg.time = "year", statistic = "mean", sites = FALSE)
}
\arguments{
\item{df}{Data frame containing the variables \code{date}, \code{site_code}, \code{value}.}

\item{avg.time}{Character vector specifying the time period over which to average the data. Option are 'year' or
'month' (to return the annual/monthly average)}

\item{statistic}{The statistic to use for averaging. Options are 'mean' or 'median'.}

\item{sites}{Logical indicating whether averaging should be grouped by site (TRUE) or data should be averaged over all sites (FALSE).}
}
\value{
Data frame of annual/monthly average concentration data
}
\description{
Average raw hourly data by year or month-year
}
