% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_trends.R
\name{average_trends}
\alias{average_trends}
\title{Compare average trend over all sites with average trend over long term sites}
\usage{
average_trends(obs, pollutant, stat = "median",
  start.date = "2000-01-01", end.date = "2017-12-31",
  avg.time = "year")
}
\arguments{
\item{obs}{A data frame of ambient pollutant concentration data. Must contain the columns: site_code, date,
value. If 'pollutant' is a pollutant ratio, the data frames of the corresponding pollutants
must be supplied as a list of data frames in the order they are given in the ratio. E.g. for
\code{pollutant = "no2/nox"}, \code{obs = list(obs.no2, obs.nox)}.}

\item{pollutant}{The pollutant of interest (character string). To calculate rolling change trend for a pollutant ratio,
separate the two pollutants with a forward slash e.g. \code{pollutant = "no2/nox"}.}

\item{stat}{The metric used to average the ambient concentration data by year. Options: "median", "mean".}

\item{start.date}{The starting date of the period of interest. Data capture to filter long term sites will be
conducted using this date as the starting date.}

\item{end.date}{The end date of the period of interest. Data capture to filter long term sites will be
conducted using this date as the ending date.}

\item{avg.time}{Resolution to average data to. Options: "year", "month", "day".}
}
\value{
A list of plots:
\itemize{
    \item \code{average.trend} - The average trend over all sites measuring during the period of interest.
    \item \code{longterm.trend} - The average trend over all long term sites measuring constantly throughout the period.
    \item \code{trends.comparison} - Comparison of the \code{average.trend} and the \code{longterm.trend}.
    \item \code{average.allsites} - The average trend for each individual sites measuring during the period of interest.
    \item \code{longterm.allsites} - The average trend for each individual long term site measuring constantly throughout the period.
 }
 Note: if there are no long term sites available for the period of interest, only the \code{average.trend} and
 \code{average.allsites} will be returned.
}
\description{
Produces plots of annual average trends in the pollutant of interest aggregated over all sites, and
over the long term sites only (i.e. sites open constantly over the period of interest).
}
\examples{
\dontrun{
average_trends(london_nox_data, pollutant = "nox", start.date = "2000-01-01", end.date = "2017-12-31")
}

}
