% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aqtrends_helpers.R
\name{calculate_pollutant_ratio}
\alias{calculate_pollutant_ratio}
\title{Function to calculate average pollutant ratios using linear model (same function as \code{average_data} but for pollutant ratios)}
\usage{
calculate_pollutant_ratio(df.list, avg.time, statistic, sites = FALSE)
}
\arguments{
\item{df.list}{List of data frames of observation data for each pollutant in the ratio. First element of list should be
data frame for the pollutant which is the numerator in the ratio (second element should be for the denominator pollutant).
Both data frames should contain the variables \code{date}, \code{site_code}, \code{value}.}

\item{avg.time}{Character vector specifying the time period over which to average the data. Option are 'year' or
'month' (to return the annual/monthly average)}

\item{statistic}{The statistic to use for averaging. Options are 'mean' or 'median'.}

\item{sites}{Logical indicating whether averaging should be grouped by site (TRUE) or data should be averaged over all sites (FALSE).}
}
\value{
Data frame of annual/monthly average concentration data for the pollutant ratio
}
\description{
Function to calculate average pollutant ratios using linear model (same function as \code{average_data} but for pollutant ratios)
}
