% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aqtrends_helpers.R
\name{check_arguments}
\alias{check_arguments}
\title{Function to check the arguments of the function and return a user-friendly error if they violate the requirements}
\usage{
check_arguments(obs = NULL, pollutant = NULL, window.width = NULL,
  stat = NULL, start.date = NULL, end.date = NULL, data.capture = NULL,
  smooth.method = NULL, parallel = NULL)
}
\arguments{
\item{obs, pollutant, window.width, stat, start.date, end.date, data.capture, smooth.method, parallel}{Pass the arguments to the
main function directly to the \code{check_arguments} function for checking.}
}
\value{
If one of the arguments violates the argument requirements, will return an informative error. Otherwise, the function
does nothing.
}
\description{
Function to check the arguments of the function and return a user-friendly error if they violate the requirements
}
