% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aqtrends_helpers.R
\name{check_date_format}
\alias{check_date_format}
\title{Check that a character string is in correct date format}
\usage{
check_date_format(date.string)
}
\arguments{
\item{date.string}{The character string of the date to be tested for whether the format is correct (\%Y-\%m-\%d)}
}
\value{
Logical value indicating whether the string is in the form YYYY-mm-dd (TRUE) or not (FALSE)
}
\description{
Check that a character string is in correct date format
}
