% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aqtrends_helpers.R
\name{data_capture}
\alias{data_capture}
\title{Function to filter observations data to only include sites with sufficient data capture over the period of interest.}
\usage{
data_capture(data, threshold, start.date = NULL, end.date = NULL)
}
\arguments{
\item{data}{A data frame of observation data with the columns: site_code, date, value (i.e. the same as
the column names in the observations table of the air_quality_data database).}

\item{threshold}{A value between 0-100 specifying the data capture threshold by which to filter the data}

\item{start.date, end.date}{A string in the format "YYYY-MM-DD" specifying the starting date and the ending date between
which to measure data capture.}
}
\value{
Data frame of observation data filtered by data capture over the defined period
}
\description{
Function to filter observations data to only include sites with sufficient data capture over the period of interest.
}
