% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aqtrends_helpers.R
\name{define_moving_windows}
\alias{define_moving_windows}
\title{Define the n moving windows of the rolling regressions}
\usage{
define_moving_windows(window.width, obs, start.date, end.date, avg.ts)
}
\arguments{
\item{window.width}{Width of the moving window}

\item{start.date, end.date}{Starting and ending dates of the entire period of analysis (e.g. "2000-01-01", "2017-12-31").}
}
\value{
Vector of starting dates of the n moving windows (in Date data type).
}
\description{
Define the n moving windows of the rolling regressions
}
