% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aqtrends_helpers.R
\name{get_longterm_sites}
\alias{get_longterm_sites}
\title{Return the site codes of the 'long term sites' (recorded measurements during every year of the period specified in the
function call)}
\usage{
get_longterm_sites(df, pollutant, start, end, resolution = "year")
}
\arguments{
\item{df}{Data frame of observation data from which to get long term sites}

\item{pollutant}{Pollutant of interest}

\item{start, end}{Starting and ending year of period of interest (e.g. 2000, 2017)}

\item{resolution}{Resolution at which to determine whether a site is long term. E.g. for a resolution of 'year', to qualify
as 'long term', a monitoring site must have measurements of the pollutant of interest in every year between \code{start}
and \code{end}. For a resolution of 'month', the site must have data measured during **each month** between \code{start}
and \code{end}.}
}
\value{
Vector of site codes of sites meeting the 'long term' criterion (measurements during every year/month/day of the
period of interest).
}
\description{
Return the site codes of the 'long term sites' (recorded measurements during every year of the period specified in the
function call)
}
