% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aqtrends_helpers.R
\name{pollutant_ratio_data_capture}
\alias{pollutant_ratio_data_capture}
\title{Calculate average pollutant ratios (with data capture constraint)}
\usage{
pollutant_ratio_data_capture(df.list, avg.time, start, end, data.capture,
  statistic = "mean", sites = FALSE)
}
\arguments{
\item{df.list}{List of data frames of observation data for each pollutant in the ratio. First element of list should be
data frame for the pollutant which is the numerator in the ratio (second element should be for the denominator pollutant).
Both data frames should contain the variables \code{date}, \code{site_code}, \code{value}.}

\item{avg.time}{Character vector specifying the time period over which to average the data. Option are 'year' or
'month' (to return the annual/monthly average)}

\item{start, end}{A string in the format "YYYY-MM-DD" specifying the starting date and the ending date between
which to measure data capture.}

\item{data.capture}{A value between 0-100 specifying the data capture threshold by which to filter the data}

\item{statistic}{The statistic to use for averaging. Options are 'mean' or 'median'.}

\item{sites}{Logical indicating whether averaging should be grouped by site (TRUE) or data should be averaged over all sites (FALSE).}
}
\value{
Data frame of annual/monthly average concentration data for the (filtered) pollutant ratio
}
\description{
Calculate average pollutant ratios (with data capture constraint)
}
