% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling_change_trend.R
\name{rolling_change_trend}
\alias{rolling_change_trend}
\title{Rolling change trends (extract true trend)}
\usage{
rolling_change_trend(obs, pollutant, window.width, stat = "median",
  start.date = "2000-01-01", end.date = "2017-12-31", data.capture = 90,
  smooth.method = "loess", parallel = FALSE)
}
\arguments{
\item{obs}{A data frame of ambient pollutant concentration data. Must contain the columns: site_code, date,
value. If 'pollutant' is a pollutant ratio, the data frames of the corresponding pollutants
must be supplied as a list of data frames in the order they are given in the ratio. E.g. for
\code{pollutant = "no2/nox"}, \code{obs = list(obs.no2, obs.nox)}.}

\item{pollutant}{The pollutant of interest (character string). To calculate rolling change trend for a pollutant ratio,
separate the two pollutants with a forward slash e.g. \code{pollutant = "no2/nox"}.}

\item{window.width}{The width of the moving window, n, over which the change in concentration is calculated (in years).}

\item{stat}{The metric (character string) used to average the ambient concentration data by year. Options: "median", "mean".}

\item{start.date, end.date}{The starting and ending dates (character string) of the period of interest over which
to calculate and plot the change trend.}

\item{data.capture}{The data capture threshold used to filter data within each moving window. During calculation of the
rolling regression, only data from sites with data capture >= \code{data.capture} over the period encapsulated by the
moving window will be included in the rolling regression.}

\item{smooth.method}{The smoothing method to use in the \code{geom_smooth} function when plotting trends.}

\item{parallel}{Logical indicating whether the rolling changes should be computed in parallel. If \code{TRUE}, the
parallelisation will be implemented using the \code{foreach} function. The number of cores used will be the total
number of cores - 1.}
}
\value{
A plot of the proxy trend (change in concentration over an N year period as a function of time).
}
\description{
Removal of distorting effect of site movement to reveal underlying trend by using changes in
concentration as a function of year as a proxy for the average trend. This method retains information about the
shape of the trend, while ignoring differences in magnitude, thus removing the leveraging effect of opening and
closing sites with extreme magnitudes.
}
\examples{
\dontrun{
rolling_change_trend(obs.nox,
pollutant = "nox",
window.width = 3,
stat = "median",
start.date = "2000-01-01", end.date = "2017-12-31")
}

}
