% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aqtrends_helpers.R
\name{rolling_worker}
\alias{rolling_worker}
\title{Worker function for calculating rolling trend over a single window (i.e. one of the moving windows)}
\usage{
rolling_worker(d1, obs, pollutant, window.width, stat, avg.ts, verbose)
}
\arguments{
\item{d1}{Date of the starting year of the moving window for which to calculate the rolling trend}

\item{pollutant}{Pollutant for which to calculate rolling trend}

\item{window.width}{Width of the moving window}

\item{stat}{Statistic used to average data (mean, median)}
}
\value{
Dataframe of rolling trend for the moving window.
}
\description{
Worker function for calculating rolling trend over a single window (i.e. one of the moving windows)
}
