% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_flux_bias.R
\name{site_flux_bias}
\alias{site_flux_bias}
\title{The difference between opening and closing sites as a function of year (v2: 05-07-2018)}
\usage{
site_flux_bias(obs, pollutant, stat = "median", smooth.method = "loess")
}
\arguments{
\item{obs}{A data frame of ambient pollutant concentration data. Must contain the columns: site_code, date,
value. If 'pollutant' is a pollutant ratio, the data frames of the corresponding pollutants
must be supplied as a list of data frames in the order they are given in the ratio. E.g. for
\code{pollutant = "no2/nox"}, \code{obs = list(obs.no2, obs.nox)}.}

\item{pollutant}{The pollutant of interest (character string). To calculate rolling change trend for a pollutant ratio,
separate the two pollutants with a forward slash e.g. \code{pollutant = "no2/nox"}.}

\item{stat}{The metric (character string) used to average the ambient concentration data by year. Options: "median", "mean".}

\item{smooth.method}{The smoothing method to use in the \code{geom_smooth} function when plotting trends.}
}
\value{
A list of plots:
\itemize{
    \item \code{difference} - a plot of the difference between the average concentration of opening sites and closing
    sites as a function of year.
    \item \code{cumulative_difference} - the cumulative sum of differences in average concentration of opening sites and
    closing sites as a function of year.
    \item \code{weighted_cumulative_difference} - the cumulative sum of differences in average concentration of opening and
    closing sites weighted by the number of opening/closing sites as a function of year.
    \item \code{compare} - a side-by-side comparison of the \code{difference}, \code{cumulative_difference} and
    \code{weighted_cumulative_difference} plots.
}
}
\description{
Detects a difference in average concentration of the pollutant of interest between sites opening
and closing in a given year (i.e. a bias towards more/less polluted locations) and any change in this bias over
time. Changes to v2:
\enumerate{
    \item Improved organisation: averaging and pollutant ratio functions wrapped in helper functions (generalised across
    all functions in the package to reduce redundancy).
    \item Plot layouts and aesthetics modified to match those of the other functions in the package.
    \item Generalised plotting function and applied over different parameters to produce the three plots (more efficient)
    \item Removed extraneous code and redundancies
    \item Included the option to specify the smoothing method when plotting - useful for data frames with few (moving) sites
    to be able to use GAM rather than loess.
    \item Slightly altered the output for greater consistency and flexibility of use.
 }
}
\examples{
\dontrun{
site_flux_bias(obs.nox, pollutant = "nox", stat = "median", smooth.method = "gam")
}

}
