% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aqtrends_helpers.R
\name{sites_open_throughout_window}
\alias{sites_open_throughout_window}
\title{Extract site codes of monitoring sites open throughout the duration of the moving window}
\usage{
sites_open_throughout_window(df, resolution = "month",
  data.capture = 1)
}
\arguments{
\item{df}{Data frame of observations data from which to extract site codes}

\item{resolution}{The time resolution for which to apply filtering. I.e. if \code{resolution} is "month", the function
will select and return monitoring sites measuring over a certain threshold of months during the moving window period.
Options: "month", "day"}

\item{data.capture}{The data capture threshold to apply to filtering the sites. A value of 1 means that the site must
have measurements for every month/day within the moving window, while a value of 0 means no filtering takes place at all.
A value of 0.9 means that the site must have measurements for at least 90% of the months/days spanned by the moving window.}
}
\value{
A vector of site codes for monitoring sites open for the entire duration of the moving window (as specified by
the filtering parameters in the arguments).
}
\description{
Extract site codes of monitoring sites open throughout the duration of the moving window
}
