% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theilsen_slope.R
\name{theilsen_slope}
\alias{theilsen_slope}
\title{Theil-Sen slope of rolling change trend}
\usage{
theilsen_slope(change_trend)
}
\arguments{
\item{change_trend}{A rolling change trend plot output from the \code{rolling_change_trend} function.}
}
\value{
Data frame with two columns: the Theil-Sen slope and the two-sided p-value
}
\description{
Calculates the Theil-Sen slope and p-value for the rolling change trend using the \code{mblm}
and \code{Kendall} packages.
}
\details{
The two-sided p-value is calculated using Mann-Kendall trend test from
the \href{https://cran.r-project.org/web/packages/Kendall/Kendall.pdf}{Kendall} package. This is used in preference
to the confidence interval calculation provided by the \code{mblm} package, which suffers from inaccuracies
due to autocorrelation of the time series data.
}
