% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aqtrends_helpers.R
\name{get_longterm_sites}
\alias{get_longterm_sites}
\title{Return the site codes of the 'long term sites' (recorded measurements during every year of the period specified in the
function call with >= \code{data.capture} % data capture).}
\usage{
get_longterm_sites(df, pollutant, start, end, resolution = "year",
  data.capture = 90)
}
\arguments{
\item{df}{Data frame of observation data from which to get long term sites}

\item{pollutant}{Pollutant of interest}

\item{start, end}{Starting and ending year of period of interest (e.g. 2000, 2017)}

\item{resolution}{Resolution at which to determine whether a site is long term. E.g. for a resolution of 'year', to qualify
as 'long term', a monitoring site must have measurements of the pollutant of interest in every year between \code{start}
and \code{end}. For a resolution of 'month', the site must have data measured during **each month** between \code{start}
and \code{end}.}

\item{data.capture}{Data capture threshold to apply to 'long term sites' data. To qualify as a long term site,
the site must (in addition to having measurements in every year/month/day of the period) have >= \code{data.capture}
% data capture over the period of analysis (\code{start} - \code{end}). Default is 90% data capture.}
}
\value{
Vector of site codes of sites meeting the 'long term' criterion (measurements during every year/month/day of the
period of interest, and sufficient data capture over period of interest).
}
\description{
Return the site codes of the 'long term sites' (recorded measurements during every year of the period specified in the
function call with >= \code{data.capture} % data capture).
}
