% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling_trends.R
\name{rolling_trends}
\alias{rolling_trends}
\title{Rolling trend plots (v2: 05-07-2018)}
\usage{
rolling_trends(obs, pollutant = "nox", window.width = c(2, 3, 5, 7, 10, 12,
  15), stat = "median", start.date = "2000-01-01",
  end.date = "2017-12-31", data.capture = 90, smooth.method = "loess",
  parallel = TRUE)
}
\arguments{
\item{obs}{A data frame of ambient pollutant concentration data. Must contain the columns: site_code, date,
value. If 'pollutant' is a pollutant ratio, the data frames of the corresponding pollutants
must be supplied as a list of data frames in the order they are given in the ratio. E.g. for
\code{pollutant = "no2/nox"}, \code{obs = list(obs.no2, obs.nox)}.}

\item{pollutant}{The pollutant of interest (character string). To calculate rolling change trend for a pollutant ratio,
separate the two pollutants with a forward slash e.g. \code{pollutant = "no2/nox"}.}

\item{window.width}{The width of the moving window, n, over which the change in concentration is calculated (in years). This
can be a vector (to compare rolling and average trends over a range of window widths) or a numeric (to return the rolling
trend for a single window width).}

\item{stat}{The metric (character string) used to average the ambient concentration data by year. Options: "median", "mean".}

\item{start.date, end.date}{The starting and ending dates (character string) of the period of interest over which
to calculate and plot the change trend.}

\item{data.capture}{The data capture threshold used to filter data within each moving window. During calculation of the
rolling regression, only data from sites with data capture >= \code{data.capture} over the period encapsulated by the
moving window will be included in the rolling regression.}

\item{smooth.method}{The smoothing method to use in the \code{geom_smooth} function when plotting trends.}

\item{parallel}{Logical indicating whether the rolling changes should be computed in parallel. If \code{TRUE}, the
parallelisation will be implemented using the \code{foreach} function. The number of cores used will be the total
number of cores - 1.}
}
\value{
A plot (or, if \code{window.width} is a vector, a list of plots labelled by the value of the window width) of the
rolling trends (left) and the average trend across the same data as was used in the rolling trends (i.e. filtered by the
\code{window.width}) (right).
}
\description{
Demonstrates a distortion of the overall trend as a consequence of the leveraging effect of opening and
closing of monitoring sites with different magnitudes over the period examined using rolling trend plots.
Changes to v2:
\enumerate{
    \item Improved organisation: averaging and pollutant ratio functions wrapped in helper functions (generalised across
    all functions in the package to reduce redundancy).
    \item Function can be implemented with/without parallelisation.
    \item Decreased redundancy: same implementation for multiple and single moving window widths.
    \item Plot layouts and aesthetics modified to match those of the other functions in the package.
    \item Extraneous options stripped out (i.e. ability to change type of rolling trend - only option is for a linear
    rolling trend i.e. rolling regression.)
    \item Plots now fitted using GAM for rolling trends and \code{smooth.method} for average trend.
    \item Slight change to method of defining moving window (implemented using date objects rather than numeric years for
    increased flexibility) in line with the implementation in the \code{change_trends_v2} function.
}
}
\examples{
\dontrun{
rolling_trends(obs.nox, pollutant = "nox",
window.width = c(2, 5, 7, 10), stat = "median",
start.date = "2000-01-01", end.date = "2017-12-31",
data.capture = 90, smooth.method = "gam",
parallel = FALSE)
}

}
