% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theilsen_slope.R
\name{theilsen_slope}
\alias{theilsen_slope}
\title{Theil-Sen slope of rolling change trend}
\usage{
theilsen_slope(change.trend, avg.time = "year")
}
\arguments{
\item{change.trend}{A rolling change trend plot output from the \code{rolling_change_trend} function.}

\item{avg.time}{Resolution for which to calculate Theil Sen slope (e.g. yearly, monthly, daily average data).}
}
\value{
Theil-Sen estimator plot and data (slope, 95% confidence intervals, significance level).
}
\description{
Calculates the Theil-Sen slope and 95% confidence intervals for the rolling change trend using the
\code{openair::TheilSen} function.
}
